Data now "encodes" to Opus and is sent via HTTP
It's a stretch to call it encodes, as it is only partially done. The
opus encoding API only encodes when the vector sent in is of lengths 120, 240, 480, 960, 1920, or 2880 per channel. It also expects an interleaved vector in case of stereo audio.

The next plan is probably to only send in interleaved stereo audio, or
use the "Frame" struct and add an in-between abstraction that handles
encoding elegantly.