Split audio into Opus provided fixed frame_size values
Splitting using vec.chunks(). The input is usually not a multiple of the
frame_size selected, so padding the last and underfilled vec with 0s.
Probably better to replace with a rolling buffer that is appended to the
front of the next packet *before* chunking.

Current approach will most certainly lead to weird auditory artifacts
because the end of a packet is NOT the end of the audio piece.
(As we are working with streaming audio).

Also reverted back to using "Frame" as it is easier to mix the audio in
the future that way.