Added a resample function using Rubato. Unused due to problems.
Problems being input and output lengths are obviously not equal in the
resampler. Input of 1029 gets output of 1120. So, considering only
allowing 48kHz music files to be inputted in the first place.

Music from YouTube and other such services could be 44.1kHz though, so
perhaps resampling live might be the correct option here. No idea!